+++ delete_loan.php
@@ -1,23 +1,26 @@
 <?php
-include("db_config.php");
+    include("db_config.php");
 	$con = mysqli_connect(DB_HOSTNAME,DB_USERNAME,DB_PASSWORD,DB_NAME);
   	if (!$con) {
       die('Could not connect: ' . mysqli_error($con));
   	}
 
-	$id=$_GET['id'];
+	if(isset($_POST['id'])){
 
-	$del = mysqli_query($con,"DELETE FROM loan  WHERE loan_no =  $id ") ;
+		$id=$_POST['id'];
 
-if ($del){
+		$del = mysqli_query($con,"DELETE FROM loan  WHERE loan_no =  $id ") ;
+
+		if ($del){
+			
+			mysqli_close($con);
+			echo "Successfully Submited";
+		}
+		else
+		{
+			echo "Not deleted.". mysqli_error($con);		
+		}
+	}
 	
-	mysqli_close($con);
-	header("Location:customer_loan.php");
-}
-else
-{
-	echo "Not deleted.". mysqli_error($con);		
-}
-
 ?>
